+++ oalg-mtx/CHANGELOG.md
@@ -11,3 +11,10 @@ and this project adheres to the
 ## [3.2.0.0] - 2025-12-14
 - extraction from oalg-base 3.1.1.0
 
+
+## [3.2.0.1] - 2025-12-23
+
+### Added
+- crTailRowsAt in the module OAlg.Entity.Matrix.Entries
+
+