+++ oalg-stc/CHANGELOG.md
@@ -11,3 +11,5 @@ and this project adheres to the
 ## [3.2.0.0] - 2025-12-14
 - extraction from oalg-base 3.1.1.0
 
+## [3.2.1.0] - 2025-12-25
+- instance for opposite field.